+++ src/device/device_impl.hpp
@@ -39,7 +39,7 @@ public:
     cl_device_id get_device_id() const;
     cl_context get_context() const;
     cl_command_queue get_command_queue() const;
-    size_t get_compute_unit_available() const; 
+    int get_compute_unit_available() const; 
     size_t estimate_compute_unit_usage(size_t work_size);
 
     void compute_unit_usage_increment(int offset);
